Add subcategories

+++ filemanager.go
@@ -7,16 +7,17 @@ import (
 	"strings"
 )
 
-func organizeFile(sourcePath, category string) error {
+func organizeFile(sourcePath, category string, subcategory string) error {
 	targetDir, exists := directories[category]
 	if !exists {
 		return fmt.Errorf("category not found: %s", category)
 	}
-	_, err := os.Stat(sourcePath)
-	if os.IsNotExist(err) {
+	if !dirExists(filepath.Join(targetDir.Path, subcategory)) && subcategory != "" {
+		return fmt.Errorf("subcategory not found")
+	}
+	targetDir.Path = filepath.Join(targetDir.Path, subcategory)
+	if !dirExists(sourcePath) {
 		return fmt.Errorf("file not found: %s", sourcePath)
-	} else if err != nil {
-		return err
 	}
 	destinationPath := filepath.Join(homeDir, targetDir.Path, filepath.Base(sourcePath))
 	if err := os.Rename(sourcePath, destinationPath); err != nil {
@@ -39,4 +40,28 @@ func isTempFile(filename string) bool {
 		}
 	}
 	return false
+}
+
+func getSubDirectories(path string) ([]string, error) {
+	entries, err := os.ReadDir(path)
+	if err != nil {
+		return nil, err
+	}
+	var subDirs []string
+	for _, entry := range entries {
+		if entry.IsDir() {
+			subDirs = append(subDirs, entry.Name())
+		}
+	}
+	return subDirs, nil
+}
+
+func dirExists(path string) bool {
+	_, err := os.Stat(path)
+	if os.IsNotExist(err) {
+		return false
+	} else if err != nil {
+		return false
+	}
+	return true
 }
